more fixes

+++ app/src/main/java/com/example/medlife/repository/MedicationDao.kt
@@ -9,7 +9,7 @@ interface MedicationDao {
     fun getAll(): Array<Medication>
 
     @Query("select * from medication where id = :id")
-    fun loadById(id: Int): Medication
+    fun loadById(id: Long): Medication
 
     //@Query("select * from medication where name = :name")
     //fun loadByName(name: String) : String
